% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccode}
\alias{ccode}
\title{Data on the countries of the world and its two digit state code}
\format{
A data frame with 249  rows and 2 variables:
\describe{
\item{Name}{Name of the country}
\item{Code}{Alpha 2 country code}

}
}
\usage{
ccode
}
\description{
Data on the countries of the world and its two digit state code
}
\keyword{datasets}
