% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverfunction.R
\name{choosevar}
\alias{choosevar}
\title{Selects variable after filtering observations}
\usage{
choosevar(frame, filterval, variable)
}
\arguments{
\item{frame}{The chosen data frame.}

\item{filterval}{The value filtered in the ID observation of the data frame.}

\item{variable}{The data frame variable to be selected.}
}
\value{
A tibble with the chosen variable with filtered values.
}
\description{
Selects a particular variable in a data frame after filtering observations to match a particular value.
}
