% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{COVID-19 in United States}
\usage{
launch_app()
}
\description{
Launches a shiny web application showcasing an overview of COVID-19 on a global scale and explores the test results as well as different dimensions of the effect of the pandemic in the United States of America.
}
