% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{months_tests}
\alias{months_tests}
\title{Data on the COVID-19 pandemic for the US and individual states.}
\format{
A data frame with 402  rows and 6 variables:
\describe{
\item{month}{Month}

\item{monthnum}{Month of the year}
\item{state_name}{Name of the state in US.}
\item{tests}{Number of tests conducted}
\item{test_positive}{Number of tests that came positive.}
\item{test_negative}{Number of tests that came negative.}

}
}
\source{
\url{https://covidtracking.com/data/download}
}
\usage{
months_tests
}
\description{
The dataset contains monthly test data on the COVID-19 pandemic for the US on individual state level .
}
\keyword{datasets}
