% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{positive_cases}
\alias{positive_cases}
\title{Data on the cumulative COVID-19 positive cases for the US and individual states.}
\format{
A data frame with 50 rows and 2 variables:
\describe{

\item{state_name}{Name of the state in US.}
\item{positive_cases}{Number of COVID-19 positive cases.}

}
}
\source{
\url{https://covidtracking.com/data/download}
}
\usage{
positive_cases
}
\description{
The dataset contains data on the cumulative COVID-19 positive cases for the US on individual state level .
}
\keyword{datasets}
